feat: implement IMU data transmission

+++ stm32/Drivers/USER/inc/time_base.hpp
@@ -8,16 +8,16 @@
 #define MILLIS_TO_MICROS(ms) ((ms) * 1000u)
 
 struct TimeBaseConfig {
-  uint32_t prescaler;     // e.g. 83 -> 1 MHz tick if TIM2CLK = 84 MHz
-  uint32_t period;        // typically 0xFFFFFFFF
-  uint32_t compensation;  // PPS capture compensation in microseconds
+  uint32_t prescaler;    // e.g. 83 -> 1 MHz tick if TIM2CLK = 84 MHz
+  uint32_t period;       // typically 0xFFFFFFFF
+  uint32_t compensation; // PPS capture compensation in microseconds
 };
 
 class TimeBase {
 public:
   using Config = TimeBaseConfig;
 
-  uint32_t Micros() const;
+  uint64_t Micros() const;
   void DelayMicros(uint32_t us) const;
 
   bool IsGpsSynchronized() const;
